Refactored the simple map test and integrated it into the test framework but the test is disabled purposely. It is a manual run test. This makes it easier to execute the test.
